refactor(#1343): split version_gate.rs into concern submodules
Mechanical split per epic #1116 (foo.rs -> foo/mod.rs + concern
submodules with re-exports; tests follow their code). No behavior change.

- version_gate/mod.rs      -- VersionGates enum + from_descriptor/from_path,
                              module docs, pub use re-exports, combined tests
- version_gate/descriptor.rs -- SsTableFormat, SsTableDescriptor + tests
- version_gate/big.rs      -- BigVersionGates (na/nb/oa allowlist) + tests
- version_gate/bti.rs      -- BtiVersionGates (da) + tests

Public API unchanged: all previously-public items re-exported from
crate::storage::sstable::version_gate::*; no call-site edits needed.
Each submodule under the 800-line src threshold (file-size gate PASS
without CQLITE_ALLOW_FILE_GROWTH).
